pyproject.toml
[project]
name = "mistral-hackathon"
version = "0.1.0"
description = "Mistral AI Voice API Proof of Concept"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=4.0.0",
    "gradio_client>=0.10.0",
    "mistralai>=1.12.4",
    "python-dotenv>=1.2.1",
]
